Move utility headers into include/utility/ for SPM public header path (Task 05)
Relocates constants.h and debug.h from library/src/utility/ to
library/src/include/utility/ and creates a new Bazel target there.
The implementation target (utility:constants) now depends on the
include target for the headers, keeping the .cpp in place.
